% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{AlphamvecOpt}
\alias{AlphamvecOpt}
\title{AlphamvecOpt}
\usage{
AlphamvecOpt(nIP, K, currentC, currentZ, alpha, mvec)
}
\arguments{
\item{nIP}{total number of interaction patterns specified by the user}

\item{K}{total number of topics specified by the user}

\item{currentC}{current state of the assignment of interaction patterns}

\item{currentZ}{current state of the assignment of topics}

\item{alpha}{Dirichlet concentration prior for topic distribution}

\item{mvec}{Dirichlet base prior for topic distribution}
}
\value{
The optimized value of the vector (= alpha * mvec)
}
\description{
Optimize the hyperparmeter vector (= alpha * mvec) given the current assignment of C and Z
}

