% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{betapartB}
\alias{betapartB}
\title{betapartB}
\usage{
betapartB(nIP, lambdai, edgeC)
}
\arguments{
\item{edgeC}{}
}
\value{

}
\description{
calculate the log of beta part used to obtain the constants for multinomial sampling of Beta
}

