% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPTM.R
\name{dinvgamma}
\alias{dinvgamma}
\title{dinvgamma}
\usage{
dinvgamma(x, shape, scale)
}
\arguments{
\item{x}{Scalar location to evaluate density}

\item{shape}{Scalar shape parameter}

\item{scale}{Scalar shape parameter}
}
\value{
evaluate the density at x
}
\description{
log of inverse-gamma pdf
}
