% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ForwardGiR}
\alias{ForwardGiR}
\title{ForwardGiR}
\usage{
ForwardGiR(Nsamp, nDocs, node, vocabulary, nIP, K, nwords, alpha, mvec, betas,
  nvec, prior.b.mean, prior.b.var, prior.delta, sigma_Q, niters,
  netstat = c("dyadic"), base.edge, base.text, generate_PP_plots = TRUE,
  seed = 1)
}
\arguments{
\item{Nsamp}{number of GiR samples to be generated}

\item{nDocs}{number of documents to be generated per each sample}

\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{vocabulary}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns specified by the user}

\item{K}{total number of topics specified by the user}

\item{nwords}{number of words in a document (fixed constant for now)}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{betas}{Dirichlet concentration prior for topic-word distribution}

\item{nvec}{Dirichlet base prior for topic-word distribution}

\item{prior.b.mean}{mean vector of b in multivariate normal distribution}

\item{prior.b.var}{covairance matrix of b in multivariate normal distribution}

\item{prior.delta}{parameter of delta in Normal prior}

\item{sigma_Q}{proposal distribution variance parameter for beta and eta}

\item{niters}{(out, n_B, n_d, burn, thinning) in the inference}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{base.edge}{artificial collection of documents to be used as initial state of history}

\item{base.text}{artificial collection of documents to be used as initial state of history}

\item{generate_PP_plots}{Logical indicating whether to draw PP plots}

\item{seed}{an integer value which controls random number generation}
}
\value{
Forward and collapsed-time forward samples and corresponding test results
}
\description{
Getting it Right test for IPTM
}
