% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPTM.R
\name{Generate}
\alias{Generate}
\title{Generate}
\usage{
Generate(D, A, psi, theta, phi, beta, eta, sigma2, X, Y, support,
  timeunit = 3600, timedist = "lognormal", prior.epsilon)
}
\arguments{
\item{D}{number of events to be generated}

\item{A}{vector of node id's (ID starting from 1)}

\item{psi}{C-length vector of interaction pattern weights}

\item{theta}{C x K matrix of pattern-topic weights}

\item{phi}{K x V matrix of topic-word weights}

\item{beta}{C x P matrix of patern-specific coefficients for recipients}

\item{eta}{C x Q matrix of patern-specific coefficients for timestamps}

\item{sigma2}{variance parameter for the timestamps}

\item{X}{an array of dimension D x A x A x P for covariates used for Gibbs measure}

\item{Y}{an array of dimension D x A x Q for covariates used for timestamps GLM}

\item{support}{support of latent recipients from Gibbs measure}

\item{timeunit}{hour (= 3600) or day (=3600*24) and so on}

\item{timedist}{lognormal or exponential (will include others)}

\item{prior.epsilon}{prior parameter in Gamma distribution}
}
\value{
generated data including (sender, recipients, timestamp)
}
\description{
Generate a collection of events according to the generative process
}
