% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GenerateDocs.PPC}
\alias{GenerateDocs.PPC}
\title{GenerateDocs.PPC}
\usage{
GenerateDocs.PPC(nDocs, node, vocab, nIP, K, alpha, mvec, beta, b, eta, delta,
  sigma2_tau, l, u, netstat, timestat, base.edge, base.text, z,
  word_type_topic_counts)
}
\arguments{
\item{nDocs}{number of documents to be generated}

\item{node}{vector of node id's (ID starting from 1)}

\item{vocab}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns}

\item{K}{total number of topics}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{beta}{Dirichlet concentration prior for topic-word distribution}

\item{b}{coefficients for recipients}

\item{eta}{coefficients for timestamps}

\item{delta}{tuning parameter for the number of recipients}

\item{sigma2_tau}{variance parameter for the timestamps}

\item{l}{topic-interaction pattern assignment}

\item{u}{inferred latent receivers}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("sender", "receiver","timeofday", "dayofweek")}

\item{base.edge}{edges before 384 hours that is used to calculate initial history of interactions}

\item{base.text}{texts corresponding to base.edge}

\item{word_type_topic_counts}{word_type_topic_counts from inference}

\item{topic_token_assignments}{matrix of topic-token assignments}
}
\value{
generated edge and text
}
\description{
Generate a collection of documents according to the generative process of IPTM using Gibbs measure
}
