% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GenerateDocs.predict}
\alias{GenerateDocs.predict}
\title{GenerateDocs.predict}
\usage{
GenerateDocs.predict(node, vocabulary, nIP, K, owords, alpha, mvec, betas, nvec,
  b, delta, currentC, netstat, initial_iJi, base.edge, base.text, currentZ,
  edge2, word_type_topic_counts, niter)
}
\arguments{
\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{vocabulary}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns specified by the user}

\item{K}{total number of topics specified by the user}

\item{owords}{observed words in a document}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{betas}{Dirichlet concentration prior for topic-word distribution}

\item{nvec}{Dirichlet base prior for topic-word distribution}

\item{b}{List of coefficients estimating the history effect for each IP}

\item{delta}{tuning parameter controlling the number of recipients}

\item{currentC}{topic-IP assignment}

\item{netstat}{which type of network statistics to use ("intercept", "dyadic", "triadic", "degree")}

\item{initial_iJi}{initial value of iJi from emprical summary of inferred latent receivers}

\item{base.edge}{edges before 384 hours that is used to calculate initial history of interactions}

\item{base.text}{texts corresponding to base.edge}

\item{currentZ}{list of topic-token assignments}

\item{edge2}{index of documents after initial history}

\item{word_type_topic_counts}{word_type_topic_counts from inference}

\item{niter}{number of iterations for Gibbs update of topic assignments and latent receivers}
}
\value{
generated edge and text, parameter b used to generate those, and base (if base == TRUE)
}
\description{
Generate a collection of documents according to the generative process of IPTM using Gibbs measure
}
