% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GenerateDocs}
\alias{GenerateDocs}
\title{GenerateDocs}
\usage{
GenerateDocs(nDocs, betas, gammas, node, vocabulary, alpha, mvec, beta, nvec,
  lambda, delta = 1, netstat = c("dyadic", "triadic", "degree"), seed = 1)
}
\arguments{
\item{nDocs}{number of documents to be generated}

\item{betas}{network statistics coefficients (beta) for each interaction pattern}

\item{gammas}{distribution of interaction patterns}

\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{vocabulary}{all vocabularies used over the corpus}

\item{alpha}{Dirichlet concentration prior for topic distribution}

\item{mvec}{Dirichlet base prior for topic distribution}

\item{beta}{Dirichlet concentration prior for topic-word distribution}

\item{nvec}{Dirichlet base prior for topic-word distribution}

\item{lambda}{parameter of speed at which sender replies, with larger values indicating faster response time}

\item{delta}{tuning parameter}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{seed}{an integer value which controls random number generation}
}
\value{
edge and text generated according to the generative process
}
\description{
Generate the documents according to the generative process
}

