% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GiR}
\alias{GiR}
\title{GiR}
\usage{
GiR(Nsamp, nDocs = 5, node = 1:4, vocabulary = c("hi", "hello", "bye",
  "mine", "what"), nIP = 2, K = 4, nwords = 4, alpha = 2,
  mvec = rep(1/4, 4), betas = 2, nvec = rep(1/5, 5),
  prior.b.mean = c(-3, rep(0, 6)), prior.b.var = diag(7), prior.eta = c(0,
  1), sigma_Q = 0.25, niters = c(1, 1, 1, 50, 0, 1), netstat = "dyadic",
  generate_PP_plots = TRUE, seed = 1)
}
\arguments{
\item{Nsamp}{number of GiR samples to be generated}

\item{nDocs}{number of documents to be generated per each sample}

\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{vocabulary}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns specified by the user}

\item{K}{total number of topics specified by the user}

\item{nwords}{number of words in a document (fixed constant for now)}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{betas}{Dirichlet concentration prior for topic-word distribution}

\item{nvec}{Dirichlet base prior for topic-word distribution}

\item{prior.b.mean}{mean vector of b in multivariate normal distribution}

\item{prior.b.var}{covairance matrix of b in multivariate normal distribution}

\item{prior.eta}{mean and variance parameter of eta in normal distribution}

\item{sigma_Q}{proposal distribution variance parameter for beta}

\item{niters}{(out, n1, n2, n3, burn, thin) in the inference}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{generate_PP_plots}{Logical indicating whether to draw PP plots}

\item{seed}{an integer value which controls random number generation}
}
\value{
Forward and Backward samples and corresponding test results
}
\description{
Getting it Right test for IPTM
}
