% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GiR_PP_Plots}
\alias{GiR_PP_Plots}
\title{GiR_PP_Plots}
\usage{
GiR_PP_Plots(Forward_stats, Backward_stats)
}
\arguments{
\item{Forward_stats}{statistics obtained from forward sampling}

\item{Backward_stats}{statistics obtained from backward sampling}
}
\value{
PP-plots for different GiR statistics of interest
}
\description{
Generate PP-plots (probability-probability plot) for Gettig it Right test
}
