% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GiR_QQ_Plots}
\alias{GiR_QQ_Plots}
\title{GiR_QQ_Plots}
\usage{
GiR_QQ_Plots(Forward_stats, Backward_stats)
}
\arguments{
\item{Forward_stats}{statistics obtained from forward sampling}

\item{Backward_stats}{statistics obtained from backward sampling}
}
\value{
QQ-plots for different GiR statistics of interest
}
\description{
Generate QQ-plots (quantile-quantile plots) for Gettig it Right test
}
