% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{history}
\alias{history}
\title{history}
\usage{
history(edge, node, when)
}
\arguments{
\item{when}{}
}
\value{

}
\description{
calculate the time difference from previous interactions to certain time 'beforetime'
}

