% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Inference}
\alias{Inference}
\title{MCMC}
\usage{
Inference(edge, node, textlist, vocabulary, nIP, K, sigma_Q, alpha, mvec, betas,
  nvec, prior.b.mean, prior.b.var, prior.delta, out, n1, n2, n3, burn, thin,
  netstat, seed, plot = FALSE)
}
\arguments{
\item{edge}{list of document information with 3 elements (element 1 sender, element 2 receiver, element 3 time in unix.time format)}

\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{textlist}{list of text (length=number of documents in total) containing the words in each document}

\item{vocabulary}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns specified by the user}

\item{K}{total number of topics specified by the user}

\item{n1}{size of first inner iteration for updates of interaction patterns C}

\item{n2}{size of second inner iteration for updates of topics K}

\item{n3}{size of third inner iteration for updates of beta B}

\item{burn}{iterations to be discarded at the beginning of beta chain}

\item{thin}{the thinning interval of beta chain}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{seed}{an integer value which controls random number generation}

\item{plot}{to plot the convergence diagnostics or not (TRUE/FALSE)}

\item{delta.B}{proposal distribution variance parameter for beta}

\item{outer}{size of outer iteration}
}
\value{
MCMC output containing IP assignment, topic assignment, and (beta, mu, delta) chain
}
\description{
Iterate Markov Chain Monte Carlo (MCMC) algorithm to sequentially update the assignments of C, Z, and B
}

