% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPTM.R
\name{Inference2}
\alias{Inference2}
\title{Inference2}
\usage{
Inference2(data, X, Y, nIP, K, V, outer, inner, burn, prior.epsilon, prior.beta,
  prior.eta, prior.sigma2, initialval = NULL, proposal.var, timeunit = 3600,
  lasttime, timedist = "lognormal")
}
\arguments{
\item{data}{list of data with 4 elements (1: sender, 2: recipient, 3: timestamp in unix.time format, 4: words)}

\item{X}{an array of dimension D x A x A x P for covariates used for Gibbs measure}

\item{Y}{an array of dimension D x A x Q for covariates used for timestamps GLM}

\item{nIP}{number of interaction patterns}

\item{K}{number of topics}

\item{V}{number of words}

\item{outer}{size of outer iterations}

\item{inner}{size of inner iteration for Metropolis-Hastings updates}

\item{burn}{size of burn-in}

\item{prior.epsilon}{prior parameter in Gamma distribution}

\item{prior.beta}{prior mean and covariance of beta in multivariate Normal distribution}

\item{prior.eta}{prior mean and covariance of eta in multivariate Normal distribution}

\item{prior.sigma2}{prior shape and scale parameter of sigma2 in inverse-Gamma distribution}

\item{initialval}{initial value of the parameters (if speficied)}

\item{proposal.var}{proposal variance for beta, eta, and sigma2}

\item{timeunit}{hour (= 3600) or day (=3600*24) and so on}

\item{lasttime}{last timestamp of the event used as initial history (in unix.time format)}

\item{timedist}{lognormal or exponential (will include others)}
}
\value{
generated data including (sender, recipients, timestamp)
}
\description{
Iterate Markov Chain Monte Carlo (MCMC) algorithm to infer the parameters
}
