% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{IPTM.inference}
\alias{IPTM.inference}
\title{IPTM.inference}
\usage{
IPTM.inference(edge, node, textlist, vocab, nIP, K, sigma.Q, alpha, mvec, beta,
  prior.b, prior.delta, prior.eta, prior.tau, Outer, Inner, burn, thin, netstat,
  timestat, optimize = FALSE, initial = NULL, timeunit = 3600,
  tz = "America/New_York")
}
\arguments{
\item{edge}{list of tie data with 3 elements (1: author, 2: recipient, 3: timestamp in unix.time format)}

\item{node}{vector of node id's (ID starting from 1)}

\item{textlist}{list of text containing the words in each document}

\item{vocab}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns}

\item{K}{total number of topics}

\item{sigma.Q}{proposal distribution variance parameter}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{beta}{Dirichlet concentration prior for topic-word distribution}

\item{prior.b}{prior mean and covariance of b in multivariate Normal distribution}

\item{prior.delta}{prior mean and variance of delta in Normal distribution}

\item{prior.eta}{prior mean and covariance of eta in multivariate Normal distribution}

\item{prior.tau}{prior shape and scale parameter of sigma_tau in inverse-Gamma distribution}

\item{Outer}{size of outer iterations}

\item{Inner}{size of inner iteration for Metropolis-Hastings updates}

\item{burn}{iterations to be discarded at the beginning of Metropolis-Hastings chains}

\item{thin}{the thinning interval of Metropolis-Hastings chains}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("sender", "receiver","timeofday", "dayofweek")}

\item{optimize}{logical to optimize alpha (Dirichlet concentration prior for document-topic distribution)}

\item{initial}{list of initial values user wants to assign including (alpha, mvec, delta, b, eta, l, z, u, sigma_tau, proposal.var1, proposal.var2)}

\item{timeunit}{hour (= 3600) or day (=3600*24) and so on}

\item{tz}{timezone such as EST and PST}
}
\value{
MCMC output containing all parameter estimates
}
\description{
Iterate Markov Chain Monte Carlo (MCMC) algorithm for the interaction-partitioned topic model
}
