% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{IPTM.PPC}
\alias{IPTM.PPC}
\title{IPTM.PPC}
\usage{
IPTM.PPC(Out, edge, node, textlist, vocab, nIP, K, netstat, timestat, inference)
}
\arguments{
\item{Out}{number of outer iterations of inference from which to generate predictions}

\item{edge}{list of document information with 3 elements (element 1 sender, element 2 receiver, element 3 time in unix.time format)}

\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{textlist}{list of text (length=number of documents in total) containing the words in each document}

\item{vocab}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns}

\item{K}{total number of topics}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("timeofday", "dayofweek")}

\item{inference}{inference result}
}
\value{
prediction output
}
\description{
posterior predictive experiments
}
