% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{logWZ}
\alias{logWZ}
\title{logWZ}
\usage{
logWZ(nIP, K, textlist2, W, zcnew, tableW, alpha, mvec, delta, nvec)
}
\arguments{
\item{nvec}{}
}
\value{

}
\description{
calculate the log of unnormalized constant (corresponding to product of Eq.13 in Section 3.2) to check the convergence
}

