% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{MCMC}
\alias{MCMC}
\title{MCMC}
\usage{
MCMC(edge, node, textlist, vocabulary, nIP, K, delta_B, outer = 200, n1 = 3,
  n2 = 3, n3 = 3300, burn = 300, thin = 3, opt = TRUE, seed = 1)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{

}

