% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Netstats}
\alias{Netstats}
\title{Netstats}
\usage{
Netstats(history, node, sender, netstat)
}
\arguments{
\item{history}{list of document information with 3 elements (sender, receiver, time)}

\item{node}{nodelist containing the ID of nodes}

\item{sender}{specific timepoint that we are calculating the time difference from}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}
}
\value{
Matrix of network statistics for specific sender and all possible receivers
}
\description{
Calculate the network statistics (dydadic, triadic, or degree) given the history of interactions
}

