% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{parupdate}
\alias{parupdate}
\title{parupdate}
\usage{
parupdate(nIP, K, currentC, currentZ, alpha, mvec)
}
\arguments{
\item{mvec}{}
}
\value{

}
\description{
parameter optimization of alpha and mvec at the same time -> need to be faster
}

