% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_betaIP}
\alias{plot_betaIP}
\title{plot_betaIP}
\usage{
plot_betaIP(MCMCchain)
}
\arguments{
\item{MCMCchain}{MCMCchain a chain obtained using MCMC function}
}
\value{
plot of the posterior summaries of beta for each interaction pattern (should click for the locator)
}
\description{
summarize the MCMC chain of network statistics (beta) for each interaction pattern, by drawing a joint boxplot
}
