% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_topic}
\alias{plot_topic}
\title{plot_topic}
\usage{
plot_topic(MCMCchain, K)
}
\arguments{
\item{MCMCchain}{MCMCchain a chain obtained using MCMC function}

\item{K}{number of topics pattern specified by the user}
}
\value{
barplot of the topic distribution
}
\description{
plot the topic distributions without considering interaction patterns
}
