% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_topicIP}
\alias{plot_topicIP}
\title{plot_topicIP}
\usage{
plot_topicIP(MCMCchain, K)
}
\arguments{
\item{MCMCchain}{MCMCchain a chain obtained using MCMC function}

\item{K}{number of topics pattern specified by the user}
}
\value{
joint barplot of the topic distribution (should click for the locator)
}
\description{
plot the topic distributions for each interaction pattern
}
