% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{PlotTopicIP}
\alias{PlotTopicIP}
\title{PlotTopicIP}
\usage{
PlotTopicIP(MCMCchain, K)
}
\arguments{
\item{MCMCchain}{MCMCchain a chain obtained using MCMC function}

\item{K}{total number of topics specified by the user}
}
\value{
Joint barplot of the topic distribution (should click for the locator)
}
\description{
Draw a barplot of the topic distributions for each interaction pattern
}

