% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPTM.R
\name{PPE}
\alias{PPE}
\title{PPE}
\usage{
PPE(data, X, Y, outer, inner, burn, prior.beta, prior.eta, prior.sigma2,
  initial = NULL, proposal.var, timeunit = 3600, lasttime, MHprop.var,
  timedist = "lognormal")
}
\arguments{
\item{data}{list of tie data with 3 elements (1: sender, 2: recipient, 3: timestamp in unix.time format)}

\item{X}{an array of dimension D x A x A x P for covariates used for Gibbs measure}

\item{Y}{an array of dimension D x A x Q for covariates used for timestamps GLM}

\item{outer}{size of outer iterations}

\item{inner}{size of inner iteration for Metropolis-Hastings updates}

\item{burn}{size of burn-in}

\item{prior.beta}{prior mean and covariance of beta in multivariate Normal distribution}

\item{prior.eta}{prior mean and covariance of eta in multivariate Normal distribution}

\item{prior.sigma2}{prior shape and scale parameter of sigma2 in inverse-Gamma distribution}

\item{initial}{initial value of the parameters (if speficied)}

\item{proposal.var}{proposal variance for beta, eta, and sigma2}

\item{timeunit}{hour (= 3600) or day (=3600*24) and so on}

\item{lasttime}{last timestamp of the event used as initial history (in unix.time format)}

\item{MHprop.var}{proposal variance for time predictions}

\item{timedist}{lognormal or exponential (will include others)}
}
\value{
generated data including (sender, recipients, timestamp)
}
\description{
Posterior predictive experiments for
}
