% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pppplot}
\alias{pppplot}
\title{PPPplot}
\usage{
pppplot(Forward_stats, Backward_stats, thin = 1)
}
\arguments{
\item{Forward_stats}{statistics obtained from forward sampling}

\item{Backward_stats}{statistics obtained from backward sampling}

\item{thin}{thinning of samples}
}
\value{
PP-plots for different GiR statistics of interest
}
\description{
Alternative way of drawing pp-plot
}
