% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPTM.R
\name{r.gibbs.measure}
\alias{r.gibbs.measure}
\title{r.gibbs.measure}
\usage{
r.gibbs.measure(lambda.i, support)
}
\arguments{
\item{lambda.i}{lambda of sender i}

\item{support}{support of Gibbs measure}
}
\value{
nsamp number of samples with each row denoting binary vector
}
\description{
Sample a binary vector from Gibbs measure
}
