% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Selected}
\alias{Selected}
\title{Selected}
\usage{
Selected(nSample, proportion)
}
\arguments{
\item{nSample}{number of samples to draw}

\item{proportion}{numeric non-negative vector of length K, specifying the probability for the K classes}
}
\value{
The vector of chosen items from the rmultinom sampling
}
\description{
Convert matrix output of rmultinom into vector of chosen items
}

