% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{table_betaIP}
\alias{table_betaIP}
\title{table_betaIP}
\usage{
table_betaIP(MCMCchain)
}
\arguments{
\item{MCMCchain}{a chain obtained using MCMC function}
}
\value{
list of table containing the posterior summaries of beta for each interaction pattern
}
\description{
summarize the MCMC chain of network statistics (beta) for each interaction pattern, by using a table
}
