% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{table_wordIP}
\alias{table_wordIP}
\title{table_wordIP}
\usage{
table_wordIP(MCMCchain, K, textlist, vocabulary)
}
\arguments{
\item{MCMCchain}{MCMCchain a chain obtained using MCMC function}

\item{K}{number of topics pattern specified by the user}

\item{textlist}{list of text (length=number of documents in total) containing the words in each document}

\item{vocabulary}{all vocabularies used over the corpus}
}
\value{
list of tables that summarize token-topic assignments of highest probabilities (topic proportion > 0.1) for each IP
}
\description{
generate the table that summarizes token-topic assignments of highest probabilities for each interaaction pattern
}
