% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{TablebetaIP}
\alias{TablebetaIP}
\title{TablebetaIP}
\usage{
TablebetaIP(MCMCchain)
}
\arguments{
\item{MCMCchain}{a chain obtained using MCMC function}
}
\value{
List of table containing the posterior summary of beta for each interaction pattern
}
\description{
Generate a table summary of the MCMC chain of network statistics coefficients (beta) for each interaction pattern
}
