% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{TopicInEqZ}
\alias{TopicInEqZ}
\title{TopicInEqZ}
\usage{
TopicInEqZ(K, currentC, currentZ, alpha, mvec, document)
}
\arguments{
\item{K}{total number of topics specified by the user}

\item{currentC}{current state of the assignment of interaction patterns}

\item{currentZ}{current state of the assignment of topics}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{document}{the id of specific edge of interest}
}
\value{
The vector of constants representing topic part of each K
}
\description{
Calculate topic-IP part of the equation used in multinomial sampling of Z
}

