% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{wordpartZ}
\alias{wordpartZ}
\title{wordpartZ}
\usage{
wordpartZ(K, textlistd, tableW, delta, nvec)
}
\arguments{
\item{nvec}{}
}
\value{

}
\description{
: calculate the log of word-topic part used to obtain the constants for multinomial sampling of K
}

