% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{betapartC}
\alias{betapartC}
\title{betapartC}
\usage{
betapartC(nIP, lambdai, specificedge)
}
\arguments{
\item{specificedge}{}
}
\value{

}
\description{
calculate the log of beta part used to obtain the constants for multinomial sampling of IP
}

