% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GenerateDocs}
\alias{GenerateDocs}
\title{GenerateDocs}
\usage{
GenerateDocs(nDocs, node, vocabulary, nIP, K, nwords, alpha, mvec, betas, nvec,
  b, delta, currentC, netstat, base.edge, base.text, seed,
  topic_token_assignments = NULL, word_type_topic_counts = NULL,
  forward = FALSE, backward_init = FALSE, backward = FALSE,
  base = FALSE)
}
\arguments{
\item{nDocs}{number of documents to be generated}

\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{vocabulary}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns specified by the user}

\item{K}{total number of topics specified by the user}

\item{nwords}{number of words in a document (fixed constant for now)}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{betas}{Dirichlet concentration prior for topic-word distribution}

\item{nvec}{Dirichlet base prior for topic-word distribution}

\item{b}{List of coefficients estimating the history effect for each IP}

\item{delta}{tuning parameter controlling the number of recipients}

\item{currentC}{topic-IP assignment}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{base.edge}{edges before 384 hours that is used to calculate initial history of interactions}

\item{base.text}{texts corresponding to base.edge}

\item{seed}{an integer value which controls random number generation}

\item{topic_token_assignments}{matrix of topic-token assignments}

\item{word_type_topic_counts}{matrix of word type - topic counts}

\item{forward}{Logigal indicating whether we are generating forward samples}

\item{backward_init}{Logigal indicating whether we are generating initial backward samples}

\item{backward}{Logigal indicating whether we are generating backward samples}

\item{base}{Logical indicating whether or not we are generating base edges (< 384)}
}
\value{
generated edge and text, parameter b used to generate those, and base (if base == TRUE)
}
\description{
Generate a collection of documents according to the generative process of IPTM
}
