% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{IPTM.PPC}
\alias{IPTM.PPC}
\title{IPTM.PPC}
\usage{
IPTM.PPC(O, edge, node, textlist, vocab, nIP, K, sigma.Q, alpha, mvec, beta,
  prior.b, prior.delta, prior.eta, prior.tau, Outer, Inner, burn, thin, netstat,
  timestat, optimize = FALSE, initial = NULL, inference)
}
\arguments{
\item{O}{number of outer iterations of inference from which to generate predictions}

\item{edge}{list of document information with 3 elements (element 1 sender, element 2 receiver, element 3 time in unix.time format)}

\item{node}{nodelist containing the ID of nodes (ID starting from 1)}

\item{textlist}{list of text (length=number of documents in total) containing the words in each document}

\item{vocab}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns}

\item{K}{total number of topics}

\item{sigma.Q}{proposal distribution variance parameter}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{beta}{Dirichlet concentration prior for topic-word distribution}

\item{prior.b}{prior mean and covariance of b in multivariate Normal distribution}

\item{prior.delta}{prior mean and variance of delta in Normal distribution}

\item{prior.eta}{prior mean and covariance of eta in multivariate Normal distribution}

\item{prior.tau}{prior shape and scale parameter of sigma2_tau in inverse-Gamma distribution}

\item{Outer}{size of outer iterations}

\item{Inner}{size of inner iteration for Metropolis-Hastings updates}

\item{burn}{iterations to be discarded at the beginning of Metropolis-Hastings chains}

\item{thin}{the thinning interval of Metropolis-Hastings chains}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("timeofday", "dayofweek")}

\item{optimize}{logical to optimize alpha (Dirichlet concentration prior for document-topic distribution)}

\item{initial}{list of initial values user wants to assign including (alpha, mvec, delta, b, eta, l, z, u, sigma2_tau, proposal.var1, proposal.var2)}

\item{inference}{inference result}
}
\value{
prediction output
}
\description{
posterior predictive experiments
}
