% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{logWZ}
\alias{logWZ}
\title{logWZ}
\usage{
logWZ(K, currentZ, textlist, tableW, alpha, mvec, betas, nvec)
}
\arguments{
\item{K}{total number of topics specified by the user}

\item{currentZ}{current state of the assignment of topics}

\item{textlist}{list of text (length=number of documents in total) containing the words in each document}

\item{tableW}{summary table of topic-word assignments}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{betas}{Dirichlet concentration prior for topic-word distribution}

\item{nvec}{Dirichlet base prior for topic-word distribution}
}
\value{
The log value of unnormalized constant
}
\description{
Calculate the log of unnormalized constant to check the convergence
}

