% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{MCMC}
\alias{MCMC}
\title{MCMC}
\usage{
MCMC(edge, node, textlist, vocabulary, nIP, K, delta_B, lambda = 0.05,
  outer = 1000, n1 = 3, n2 = 3, n3 = 5500, burn = 500, thin = 5,
  seed = 1, plot = TRUE)
}
\arguments{
\item{edge}{edgelist in the form of matrix with 3 columns (col1:sender, col2:receiver, col3=time in unix.time format)}

\item{node}{nodelist containing the ID of nodes (ID's starting from 1)}

\item{textlist}{list of text (length=number of documents in total) containing the words in each document}

\item{vocabulary}{all vocabularies used over the corpus}

\item{nIP}{number of interaction patterns specified by the user}

\item{K}{number of topics pattern specified by the user}

\item{delta_B}{proposal distribution variance parameter for beta}

\item{lambda}{parameter of speed at which sender replies, with larger values indicating faster response time}

\item{outer}{size of outer iteration}

\item{n1}{size of first inner iteration for updates of interaction patterns C}

\item{n2}{size of second inner iteration for updates of topics K}

\item{n3}{size of third inner iteration for updates of betas B}

\item{burn}{iterations to be discarded at the beginning of the chain}

\item{thin}{the thinning interval}

\item{seed}{an integer value which controls random number generation}

\item{plot}{to plot the convergence diagnostics or not (TRUE/FALSE)}
}
\value{
MCMC output containing IP assignment, topic assignment, beta, the plots to check the convergence
}
\description{
...
}
