% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Netstats}
\alias{Netstats}
\title{Netstats}
\usage{
Netstats(historyIP, node, sender, netstat)
}
\arguments{
\item{historyIP}{list containing the weighted history of interactions for each IP}

\item{node}{nodelist containing the ID of nodes}

\item{sender}{specific timepoint that we are calculating the time difference from}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{nIP}{total number of interaction patterns specified by the user}
}
\value{
list of network statistics for specific sender and all possible receivers for each IP
}
\description{
Calculate the network statistics (dydadic, triadic, or degree) given the history of interactions
}

