% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{r.gibbs.measure}
\alias{r.gibbs.measure}
\title{r.gibbs.measure}
\usage{
r.gibbs.measure(nsamp, lambda.i, delta, support)
}
\arguments{
\item{nsamp}{the number of binary vector samples to draw}

\item{lambda.i}{a vector of coefficients according to which each element}

\item{delta}{a positive real valued parameter that controls the density penalty}

\item{support}{support of Gibbs measure}
}
\value{
nsamp number of samples with each row denoting binary vector
}
\description{
List out the support of Gibbs measure
}
