% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{selected}
\alias{selected}
\title{selected}
\usage{
selected(samples, proportions)
}
\arguments{
\item{samples}{number of samples to draw}

\item{proportions}{numeric non-negative vector of length K, specifying the probability for the K classes}
}
\value{
the chosen category from the rmultinom
}
\description{
find out the chosen category from the multinomial distribution
}
