% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{selected}
\alias{selected}
\title{selected}
\usage{
selected(samples, proportions)
}
\arguments{
\item{proportions}{}
}
\value{

}
\description{
find out the chosen category from the multinomial distribution
}

