% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{topicpartC}
\alias{topicpartC}
\title{topicpartC}
\usage{
topicpartC(nIP, currentZ, zc, alpha, mvec, document)
}
\arguments{
\item{document}{}
}
\value{

}
\description{
calculate the log of topic-IP part used to obtain the constants for multinomial sampling of IP
}

