% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{topicpartZ}
\alias{topicpartZ}
\title{topicpartZ}
\usage{
topicpartZ(currentIP, zcnew, alpha, mvec)
}
\arguments{
\item{mvec}{}
}
\value{

}
\description{
calculate the log of topic-IP part used to obtain the constants for multinomial sampling of K
}

